Refactored Butler to use Neuro.Model to store _accessors.
Then extended Neuro.Model with it with custom accessor abilities.

This ties in with deep set/get refactoring. Custom accessor's should be able to handle them as well.

Need to update paths to the right files. src/model/main.js will be the one that src/main.js pulls in. So Collection needs to pull in Model from src/model/main.js